Add audio capture for browser source
Relies on two cef commits:
58e1149 : Add ability to capture audio output to buffer via cef handler
1501768 : Mute audio in the browser (issue #1806)

The audio from browser source is now muted to desktop.
It can be captured and monitored.
The PR is compatible with multiple audio sources in same browser page.

Main author: Osiris.
Other authors: Andersama, pkv & WizardCM

In memoriam.
Thanks for all that you did for this project. We deeply miss the
opportunity to keep working with you. - andersama, pkv, WizardCM

Co-authored-by: Michel Snippe <michel@sokar.nl>
Co-authored-by: Andersama <anderson.john.alexander@gmail.com>
Co-authored-by: pkv <pkv.stream@gmail.com>
Co-authored-by: WizardCM <git@wizardcm.com>